Fix remaining tests; change to float...

+++ src/main/java/com/o19s/es/ltr/ranker/normalizer/FeatureNormalizer.java
@@ -3,7 +3,7 @@ package com.o19s.es.ltr.ranker.normalizer;
 import com.o19s.es.ltr.feature.store.FeatureNormalizerFactory;
 
 public interface FeatureNormalizer {
-    double normalize(double value);
+    float normalize(float value);
 
     String featureName();
